add: cloneWithUnwrapProxyに包む
